Add HtmlUnit test
So what does it take to write the same test for our `CfpController` but
actually exercising the HTML form? Well, not much. If we add htmlunit to
our application, a `WebClient` will be auto-configured for us and its
lifecycle will be managed automatically.

All we have to do now is interact with the actual form.